requirements.txt
matplotlib==3.8.0
numpy==1.26.0
pandas==2.1.3
PyYAML==6.0.1
scikit-learn==1.3.1

